rework rosetta unit files
Only use one unit to make the flow simpler and move most of the logic
into the script to make things simpler to write and only ignore error
for the mount error (if rosetta was disabled in podman machine).
Add system conditions to only ever run on apple arm64 (needs testing as
I do not have apple hardware). With that we can enable the service by
default.
Start directly after systemd-binfmt.service, the multi-user.target can
be to late as it is also used to start containers.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>