Fix setuid binaries for x86_64 executables.
In case of binfmt, the kernel by default uses the interpreter's permissions instead of the actual binary. That is, x86_64 binaries on arm64 with setuid bit usually fail to run. Details: https://stackoverflow.com/questions/75954301/using-sudo-in-podman-with-qemu-architecture-emulation-leads-to-sudo-effective-u

This change updates the rosetta registration to force the usage of the subject binary instead of the interpreter in the podman machine.

Signed-off-by: Gergo Huszty <gergo.huszty@ibm.com>