First pass on machine image tests
These tests can be used to verify that a new, generated disk image is
valid.  Right now, it is just some framework and series of very basic
tests (the tests I manually use to verify an image).  Tests can be added
to tighten things up for sure, just remember these are tests to *verify
the image* and not validate Podman machine.

This suite also makes some assumptions, like:
  * it the only one running machine tests
  * you have fed it the path to the image or used MACHINE_IMAGE_PATH
    environment variable
  * only runs on Linux right now

JIRA #2175

Signed-off-by: Brent Baude <bbaude@redhat.com>