TMT: generate dev and release artifacts
This patch allows building dev / release artifacts with the help of a
few envvars in `podman-rpm-info.sh`. If `PODMAN_RPM_TYPE` is `dev`, it'll
fetch podman and other relevant rpms from the `podman-next` copr. Else,
it'll fetch the podman rpm from koji with version-release values as
mentioned in `podman-rpm-info-vars.sh`.

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>