add pr comment with artifact download URLs
So user have easy access to the new URLs without having to click through
the cirrus web interface. As the action triggers on check_suite this
will only work once merged into main. We cannot test this on this PR so
I am not sure if it will work the first time around.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>