build: better version check
If PODMAN_RPM_TYPE is set to dev the Containerfile will pull from the
podman-next copr. This if fine for our podman x.y+1 images but for the
releases we need the proper version. Add some validation and only allow
the two values "dev" and "release" so it is more clear what version is
build.

And also bump the version to 5.4.0-dev as this is what the copr should
pull now.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>