print-artifacts-urls.sh: update output format
Create one table with x86_64 and aarch64 and create hyperlinks as there
is no need to dump the full URL in a github comment. Users can just
click on the name which makes the output much more human readable.

Also include the link to the cirrus build for easy access the the CI
status.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>