use github.com/coreos/custom-coreos-disk-images
Use the new build script to generate the images.
Given the new script creates a zip and gz compressed files we must
uncompress them first in order to compress them again with zstd. This
makes the script much more complicated than it should be but at least it
works for now. Given that the script is also slower so we have to bump
the cirrus timeout.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>