disable packit
We use cirrus for CI as packit doesn't support testing on macos/windows
yet, has no concept of task dependecies and secrets management which is
needed to push the quay.io images securely.

So the packit task for now is just unnecessary duplication and thus
should be disabled until we can use packit/tmt for everything.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>