New VM images
from https://github.com/containers/automation_images/pull/396

The images should include the required rpms so we this should speed up
the tests as we do not have to pull them and should not unexpectly break
the build.

Also remove the extra timeout, it is no longer needed with the build
script speed up from commit 8ead68e.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>