Fix ssh auth errors when starting the machine
On macos we often seem to hit this error when starting the machine:
machine did not transition into running state: ssh error: ssh: handshake failed: EOF

As it turns out the problem is caused by the new PerSourcePenalties sshd
feature, after 5 failed log ins it locks us out. As podman machine only
runs locally we do not need that extra security and can just disable it.
It is unclear why our early login attempts are counted towards failed
attempts as we always use the same ssh key but this here fixes the
problem. And for now we like to get this in to get the podman CI working
again.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>