Add ansible to podman-machine-os
add ansible to the machine image so users can execute a playbook once
the machine is booted.  Users have been asking for a way to automaticaly
configure a machine with various options, files (like certificates), and
configuration tweaks.  Ansible is an excellent tool to do this in the
machine.

See also containers/podman#25043

Signed-off-by: Brent Baude <bbaude@redhat.com>