cirrus.yml: fix broken persistent_worker filter
We must specify the right labels so we are only scheduled on the correct
worker. This only worked because we only ever had one type of worker.
However since I gave Adrian access to our test pool he has registered
other workers with a different purpose label as I requested.
But since the yaml here was wrong cirrus did not use the purpose label
to select the right worker node. So we accidentally run on on Adrians
nodes.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>