Containerfile: run ostree container commit after rpm-ostree
From the docs[1]: In a container build, it’s a current best practice
to invoke this at the end of each RUN instruction.

Thus we should add it here, I did not add it for other RUN's as those
are simple file operations that should not effect other things really.

[1] https://coreos.github.io/rpm-ostree/container/#using-ostree-container-commit

Signed-off-by: Paul Holzinger <pholzing@redhat.com>