Release RPM: Fetch from Podman upstream PR
This commit will allow fetching the podman rpm from the upstream Podman release
PR's (or any PR for that matter) Packit copr job.

At this point, the upstream release PR is assumed to already have successful copr rpms.

We do not need a new tag pushed to podman upstream in order for this
to work.

The copr created by Packit on the Podman release PR will be preserved
indefinitely.

Note: For now, the PODMAN_VERSION var will also have to be updated
manually on release PRs.

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>