verify: use our testimage
We have no control over the docker image so it could break us at
unexpected times. More importantly docker hub has strict rate limts so
it is possible that it will cause flakes in our CI runners.

quay.io is also known to be flaky but it should be better than running
in rate limits. If flakes become an issue we should consider keeping a
copy of the image in the repo and load it instead of pulling.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit 0b17d4063e8d0b4c8b6ff4257cf2186ba641ae40)
Signed-off-by: Paul Holzinger <pholzing@redhat.com>