Cirrus: Automate release
Create release and upload built images on tag. Re-label podman release PR if needed. Also, only push images on tags, or if on main

Note that we use two different tokens for the release creation and the PR label. This is so we can give each token more fine grained access to only the repo they need to modify.

Signed-off-by: Ashley Cui <acui@redhat.com>