Containerfile: correctly disable systemd-resolved
An rpm-ostree update broke our builds as resolv.conf in the machine was
pointing the the systemd-resolved file and because we disabled the
service it was not working and NetworkManager didn't populate the file
either because of it.

The reason this broke is because of an ostree-ext change[1] which trest
the empty resolv.conf file we used in a spical way and removed it which
then allowed the tmpfiles.d rule from the pakcage to recreate the
symlink. The proper fix is to remove the package instead as pointed out
by Colin.

[1] https://github.com/bootc-dev/bootc/pull/1167

Signed-off-by: Paul Holzinger <pholzing@redhat.com>