add email notifications for failed runs on branches
So that we now when something is not right outside of PRs.
We are reusing the same action as it uses the same trigger and we must
inspect the cirrus ci_retrospective anyway to get the data we want.
So rename the action to be more descriptive and add steps to send the
email on failures.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>