cirrus: rework build setup
Now that the verify suite takes over 3 minutes it is sbetter to split it
out to another machine. With that windows/macos testing can start right
away and don't depend on the linux verify to successed before which
should help bring the total time down hopefully.

Also it means we don't have to use the expensive metal instances for the
full build which could not make use of the many cores anyway.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>