image: fix podman-user-wait-network-online.service
The main issue is that if network-online.target is not active as root
then polling for it as rootless will always fail. By default on coreos
nothing pulls in network-online.target so the service never is
triggered. This then causes problem for the rootless quadlet units.

Because we know we are always using NetworkManager we can just use their
online command directly from a rootless user to fix this.

Fixes: containers/podman#26134

Signed-off-by: Paul Holzinger <pholzing@redhat.com>