emulators should preserve argv0
Ensure we set the P flag for the binfmt setup which seems to be
supported for both qemu and rosetta. With that option set they actually
keep argv[0] instead of replacing it with the executable path.

That is required to make any application work that relies on argv[0],
such as a login shell passing.

Fixes: containers/podman#26656

Signed-off-by: Paul Holzinger <pholzing@redhat.com>