rework artifact name logic
The apple x86 removal broke the push as the gather script was not
updated.

To try to fix this do a even bigger rework do avoid looping everywhere
again based on the arch and add helper function to get all the artifact
names.

Also update the cirrus print script to make it more robust.
With these changes it should be more trival to add more providers or
even arches without much changes.

Fixes: b0fad2b945 ("Remove Intel Apple builds and artifacts")

Signed-off-by: Paul Holzinger <pholzing@redhat.com>